pipeline/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
newspaper3k>=0.2.8
lxml[html_clean]>=4.9.3
scikit-learn>=1.3.0
numpy>=1.24.0
anthropic>=0.25.0
